Fixed ProtoBuf and Added PacketSession
Fixed problem list
- If the sendbuffer is full, it reset the usedSize position to 0 without an error now.
- The packet type of a message can be expressed more widely.
CustomBuffer: within 2 bytes (ushorts)
ProtobufWrapper: within 2 bytes (ushorts)
Protobuf: within 4 bytes (int)
- Protobuf uses the message size as int type and the size is serialized with base128 encoding (varint).
(The size of the length and type size of Protobuf was fixed value, 1. It caused critical problems when deserializing the multiple data or long-length data.

Added list
- The content codes can use PacketSession of each Buffers(CustomBuffer, Protobuf, ProtobufWrapper) to create a Session class now.
- Reading and writing uint value is now available manually using Utils.Base128Encoding.

Modified
- ProtobufWrapper serializes the message to : [size, 2][packetType, 2][message]
The size is the sum of packetType Length(2) and the calculated size of the message.
- Protobuf serialized the message to :
[size, (varint)][message]
The size is the calculated size of the message and the message contains the Packet Type value in the proto tag=1.
Note: The PacketType value must be the first tag in proto.