JobTimer, JobQueue
JobTimer: JobTimer is a singleton class that schedules Actions based on specified ticks. It utilizes a priority queue to store Actions and executes them when the corresponding tick is reached. This class is designed for prioritized execution of specific tasks.

JobQueue: JobQueue is a multi-threaded class that handles the addition and execution of Actions or IJobs. It manages and executes Jobs added by multiple threads using locks to ensure thread safety.